test: Participations 자동차 이름들 단위 테스트
